requirements.txt
fastapi[all]
uvicorn
sqlalchemy
python-decouple

python-jose[cryptography]
passlib[bcrypt]